pyproject.toml
[project]
name = "feedgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "click>=8.2.1",
    "feedgenerator>=2.2.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]

[project.scripts]
feedgen = "feedgen.cli.main:cli"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
